remove distant normal normal conj test
Summary: Removes the distant normal normal test. Almost all of the inference algorithms fail this test, even external implementations that are verified to be correct. It's not a good test since it's a (very) misspecified model and there's no reason to expect that algorithms to pass a strict hypothesis test on such a model.

Differential Revision: D28528998

fbshipit-source-id: 76438377c2c76f466ae885a36623618cec08e5e3