Fix cut-n-pasted error in AST construction
Summary:
There is a small mistake that has been cut-n-pasted into a number of places in the AST rewriter.

The issue is that `new_name` is not a string, it is a `Name` object, but it is being used in a context in which a string is expected. That is, we are constructing

    whatever=Name(id=new_name, ctx=Load())`

but `new_name` is *already* a `Name` object. It should just be `whatever=new_name`.

Why was this bug undetected by the single assignment rewriter test cases?  Because those test cases use the output of `astor.to_source` to check for correctness, but `astor` does not check that the AST is *sensible*.  It just converts the `id` to a string, and if the `id` is incorrectly another `Name` object, it converts that to a string in turn.

But when you pass a malformed AST to `compile`, that raises an error saying that `id` is required to be a string.  For example, if you attempt to compile a random variable method with a line such as

    a[b][c] = d

then when it attempts to rewrite as

    x = a[b]
    x[c] = d

the `x` in the second line is `Name(id=Name(id="x"))` instead of `Name(id="x")`, and compilation of the single assignment form then crashes.

Reviewed By: wtaha

Differential Revision: D28545987

fbshipit-source-id: aae91498fbc3215554bd1dee0f4144a307c827a8