Implement runtime support for models with inversion operators (#843)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/843

BMG does not support the bit inversion operator `~` but it is a bad user experience to have the compiler crash during graph accumulation with a type error should you use it.

We now accumulate an `InvertNode` into the graph, and then give an error when checking whether the graph has an equivalent BMG graph.

I'm going to be adding a lot of these unsupported nodes and I do not want to implement size and support computations for each. I want to eventually refactor the nodes so that the size and support computations are isolated to their own module rather than spread out across the nodes, as we did for type checking.

Therefore I've made size and support no longer abstract methods; the base implementation throws an exception.  We will revisit this code when we get conditional stochastic control flows working.

Reviewed By: wtaha

Differential Revision: D28424267

fbshipit-source-id: 76d97a14b86b0c8d8221572f7e44d179fa08c726