Add error detection for more unsupported binary operators (#845)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/845

We have no support in BMG for the `& ^ | << >> % //` operators. Previous to this diff, attempting to compile a model which contained one of these operators with a stochastic operand crashed the accumulator.

That's a bad user experience. We now accumulate them into the graph and then during the graph analysis stage produce a more helpful error message.

We still need to handle the `@` (matrix multiply), `is`, `is not`, `in` and `not in` operators to produce appropriate error messages. That will happen in upcoming diffs.

Reviewed By: wtaha

Differential Revision: D28428406

fbshipit-source-id: 12a9e5a2974054dcd2d2071965b7f4b4ae5942d9