Initial global inference design (with HMC and NUTS) (#816)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/816

Related doc on global inference: https://fb.quip.com/rqrfA5GYzBmP

This diff adds a global inference framework with two algorithms (HMC + its variant, NUTS). The content of the entire diff sits inside of `beanmachine/ppl/experimental/global_inference` because it contains some rather big changes comparing to our old single-site framework, and in the next few weeks or so there might still be quite a few framework-level changes on global inference, so I want to make sure that people are okay with it before merging it to `beanmachine/ppl/inference`.

Both of the new algorithms can be directly imported from `beanmachine.ppl` namespace. i.e., if you have `import beanmachine.ppl as bm` at the top of a file, then you can use `bm.GlobalHamiltonianMonteCarlo` and `bm.GlobalNoUTurnSampler` to access the two algorithms.

The use of "world" in the global inference framework is almost identical to it is in MiniBM. In single-site inference, `World` is a singleton object. Proposed values are kept in a stack of diffs, and are either merged to the world or discarded depending on whether the new state is accepted or not. In global inference, instead of modifying the current state in-place, the proposer returns a new world object each iteration.

So we no longer need the Diff logics in single site worlds, nor do we need the functionality to compute the change in log prob as a result of a change in a single node while fixing the rest of the world at the old state (we can just compute the joint log prob of the entire world). The end result is a new world class -- here I call it `SimpleWorld` because the name `World` is already taken and I am very bad at naming things. Feel free to suggest another name for the class if you have a better idea :)

I tried follow the old design, where the inference class themselves are stateless (i.e. not bound to any model), whereas the proposers are more or less stateful. I also included the `sampler` API in this diff, which allow users to generate samples interactively ([see this old doc for some descriptions](https://fb.quip.com/Nb3TATvObVlD)). You can find some demo of its potential usage in the unit test.

Reviewed By: neerajprad

Differential Revision: D28215033

fbshipit-source-id: 7f741c314c92dd01fde50a768ec2366ab4f24490