Fix a bug in global NUT (introduced accidentally when updating previous diff)
Summary:
In one of the update in the my diff (D28215033 (https://github.com/facebookincubator/beanmachine/commit/e988549b1846f7cb1045375ab2306fcce47fbf12)) I modified `SimpleWorld`'s API so that `__getitem__` returns a variable in the constrained space, but I forgot to update the API call on the inference side.

I tried benchmark with a logistic regression model but that didn't catch the issue because it does not require transform. :(

Differential Revision: D28645807

fbshipit-source-id: 9d8f38d55ae8b0821cde9ad2aa2351616336f831