Fix a bug in global NUT (introduced accidentally when updating previous diff) (#847)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/847

In one of the update in the my diff (D28215033 (https://github.com/facebookincubator/beanmachine/commit/e988549b1846f7cb1045375ab2306fcce47fbf12)) I modified `SimpleWorld`'s API so that `__getitem__` returns a variable in the constrained space, but I forgot to update the API call on the inference side.

I tried benchmark with a logistic regression model but that didn't catch the issue because it does not require transform. :(

Reviewed By: jpchen

Differential Revision: D28645807

fbshipit-source-id: 98b103d4bc133f8f513860c666d2b99f68ab0c5c