Test case for newest CLARA model (#848)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/848

Now that we have support in BMG and the compiler for column-indexing, vectorized log-sum-exp and construction of matrices we can compile a version of the latest CLARA model which uses these features.

Note that the compiler does not yet support tuple indexing; use `[x][y]` instead of `[x, y]`.  And the compiler still does not support `+=` operators.

Reviewed By: kshah1997

Differential Revision: D28527163

fbshipit-source-id: 4598882a6709b898586b18b3261453b203ba13d8