BMG now allows queries on constants (#851)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/851

Before this change, BMG would throw an exception if you asked it to query a constant value. Though it is somewhat odd to ask "what is the value of this constant when given these observations?" there are situations where it might arise. For instance:

* Our own internal testing non-inference aspects of BMG APIs.

* A compiler producing a BMG graph as its output might be compiling a model where the developer has replaced a stochastic computation with a constant temporarily for testing purposes. It is onerous on the compiler to ensure that a query is never a constant.

* A compiler might have an optimization which deduces that a particular quantity that appears to be stochastic is in fact a constant. (Multiply by zero, stochastic IF with same constant on both sides, sum of x and its complement, and so on.)  It might be useful for the compiler to warn the user that it is doing so, but it should be allowed to do so.

* And so on.

We now allow a query on an operator or a constant. We still do not allow a query on a distribution or a factor, as doing so is not meaningful.

Reviewed By: wtaha

Differential Revision: D28714333

fbshipit-source-id: d9ac62e8d520f08ea6eaaa5274b980a14df74d05