Add to_graph function to BMGInference
Summary:
BMGInference now has the ability to take in observations and queries and directly return the equivalent BMG graph.

When you call `infer` on the resulting `Graph`, you'll get back a list of results where each queried random variable is associated with a particular element in that list. I therefore also return a map which takes the `RVIdentifier` of a query and tells you what the index in the result set is for that query. Note that it is NOT necessarily the same as the index in the original `queries` parameter.

Differential Revision: D28731773

fbshipit-source-id: 59759b5fb99ed27793676b9c0cef7665cd03f078