Factor graph performance testing flag was not read correctly (#850)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/850

We have an undocumented flag for doing performance testing of factor graphs. My intention was to have the problem fixer detect this flag on the graph builder and turn on generation of factors, but I failed to complete the necessary refactoring.

The consequence of the error, unfortunately, was that the factor graphs were ONLY generated when `to_dot` was called, which is of course what I use to test the flag.  Factors were not generated on calling `to_python`, `to_cpp`, or `infer`.

Reviewed By: wtaha, kshah1997

Differential Revision: D28727102

fbshipit-source-id: 967db25f3cec26df374a5f2582ee1bf245c76b3d