Add to_graph function to BMGInference (#852)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/852

BMGInference now has the ability to take in observations and queries and directly return the equivalent BMG graph.

When you call `infer` on the resulting `Graph`, you'll get back a list of results where each queried random variable is associated with a particular element in that list. I therefore also return a map which takes the `RVIdentifier` of a query and tells you what the index in the result set is for that query. Note that it is NOT necessarily the same as the index in the original `queries` parameter.

Reviewed By: wtaha

Differential Revision: D28731773

fbshipit-source-id: 29003fe8c240e5cb42d2ba2d25251e30cefe8843