Add TO_NEG_REAL support to compiler
Summary:
As noted in the previous diff, just as we sometimes have need to coerce a real or positive real to probability, so too do we need to coerce reals to negative reals -- because negative reals are just log probabilities.

I've added support to the compiler to detect situations where a negative real is required but a real is supplied; we then insert a TO_NEG_REAL operator which checks at runtime during inference that the range condition is met.

Differential Revision: D28761309

fbshipit-source-id: d9e188dad6d62f9ae73a2652da7380dfd10a6951