Implement TO_NEG_REAL BMG graph node (#854)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/854

In BMG we have a TO_PROBABILITY node for cases when we have a real or positive real quantity, we know from some rule of algebra that it is actually a probability, but we cannot prove it in the type system.  We insert it as needed in graphs when compiling models.

Log probabilities are negative reals, and it is also necessary in some situations to say "we have a real number but we know algebraically that it is actually negative". This diff implements a node which acts like TO_PROBABILITY but for negative reals.

In an upcoming diff I will modify the compiler to generate this node when necessary to convert a real to a negative real.

Reviewed By: yucenli

Differential Revision: D28759914

fbshipit-source-id: e6be022f523b712dacef07fe100fa35a45d496ba