Add TO_NEG_REAL support to compiler (#855)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/855

As noted in the previous diff, just as we sometimes have need to coerce a real or positive real to probability, so too do we need to coerce reals to negative reals -- because negative reals are just log probabilities.

I've added support to the compiler to detect situations where a negative real is required but a real is supplied; we then insert a TO_NEG_REAL operator which checks at runtime during inference that the range condition is met.

Reviewed By: wtaha

Differential Revision: D28761309

fbshipit-source-id: 65f2f4caa2b3eb9cf477ac3edb79cfc46d310e0b