remove distant normal normal conj test (#840)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/840

Removes the distant normal normal test. Almost all of the inference algorithms fail this test, even external implementations that are verified to be correct. It's not a good test since it's a (very) misspecified model and there's no reason to expect that algorithms to pass a strict hypothesis test on such a model.

Reviewed By: wtaha

Differential Revision: D28528998

fbshipit-source-id: 9ffce002f2a50f1816200e992cc09de112b74cfa