Tutorial on Sparse Logistic Regression as Beanstalk test case
Summary: This diff is part of a stack aimed at assessing and advancing Beanstalk's readiness for OSS release. This particular diff captures the code of the tutorial on Sparse Logistic Regreation for use as a Beanstalk test case. The test case is effective at catching on bug indicated in the comment at the top of the file. Diffs higher up in this stack will address this issue.

Differential Revision: D28127251

fbshipit-source-id: 1838519445e2bb39dd2621e6963b834e7adbbdea