Tutorial on Bayesian Logistic Regression as new Beanstalk test case
Summary:
This is the first diff in a task aimed at creating Beanstalk end-to-end test cases out of the current set of tutorials in the Bean Machine documentation.

This is a new set of tests, including one active and one disabled. The goal is to follow with diffs that provide the functionalty needed to run the disabled tests. A TODO item at the start of the file documents the specific issue enountered so far.

Differential Revision: D28067624

fbshipit-source-id: 8198f13829639ee81ba8a777ebe8a59d4a1a24cc