Tutorial on Neals Funnel as a Beanstalk test case
Summary: This diff is part of a stack aimed at assessing and advancing Beanstalk readiness for OSS. This particular diff captures the code of the tutorial on Neals Funnel as a Beanstalk test case. The disabled test raises an error that is captured in the header comment of the file. The plan is to address this issue further up in this task.

Reviewed By: ericlippert

Differential Revision: D28128787

fbshipit-source-id: 567d1cb5ebc939547b80b91d25e47e097e171ec1