Remove/fix incorrect assertions
Summary: Assertions are currently disabled in BMG, which means that as code has been edited, errors have crept into existing assertions that were not flagged by the compiler.  I mean to eventually turn assertions back on, so the first step is to eliminate or fix incorrect assertions. See comments below for details.

Differential Revision: D29040345

fbshipit-source-id: 8ed5daaefa23fda2b44ebf8c6cf05d5493cc58e9