Add test showing multiple queries on a non-constant node is broken (#871)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/871

It is rare but possible to end up with two query nodes pointing to the same non-constant node in the graph, and BMG crashes when this happens. We need to come up with some way to avoid this crash.

Two possibilities:

* Make BMG idempotent when querying a node twice. There is no requirement that this be an error. BMG returns the index into the sample associated with this query, and that can be the same no matter how many times you ask to query a node!
* Keep it an error in BMG but avoid the error condition. Write another pass which examines all the queries to find pairs of queries that refer to the same node, and "merge" the two queries so that both are associated with the two RVIDs.

The former solution seems better to me but I will give it some thought.

Reviewed By: wtaha

Differential Revision: D28815771

fbshipit-source-id: aa8e6b9ad35cc479d18407e4b6cdb604108b0097