Make adding a query to BMG idempotent. (#872)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/872

As noted in the previous diff, we need to somehow handle the situation where the compiler produces a graph with two queries on the same node; either BMG has to allow that, or the compiler needs to deduplicate the query nodes while continuing to track which queries are associated with which RVIDs.

There is no particular downside I'm aware of to making BMG allow multiple queries on the same node; the query ID is just the index into the sample vector returned by BMG, so asking for the same node to be queried twice can be implemented by returning the same index again.

Reviewed By: wtaha

Differential Revision: D28817397

fbshipit-source-id: 54a879a8fcad3c60e0b9de6f3baf239df5b25ab7