Error on operations involving unsupported constant inputs
Summary:
If a constant value makes it into the accumulated graph then either it was queried, or it was used in a calculation with stochastic inputs. We should give an error if that constant value is a tensor that cannot be represented in BMG; BMG can only do 1-D and 2-D tensors.

We achieve this by adding a case to the unsupported operation error detection pass. The error message is re-used from the error used for invalid operations, and reads poorly as a result. I will update the error message in a later diff to make it more clear.

Reviewed By: wtaha

Differential Revision: D28840962

fbshipit-source-id: 8a1bb468a442eecd38acb997159752f6800fccb4