Compiler now allows queries on constants when generating code or doing inference
Summary:
Prior to recent changes, BMG did not support querying a constant. But querying a constant is (1) legal in Bean Machine, and (2) possibly useful in some scenarios, and (3) it is irritating to have to work around this restriction.  I have added the ability to do so in BMG in a prior diff.

This diff removes the code from the compiler which did an end run around this restriction. We no longer suppress queries on constants from code generation or graph generation.

Doing so has exposed an interesting bug in the code and graph generators, and the fix is likely to be somewhat complicated and might take a few diffs to complete. I am therefore going to leave the code in its current somewhat broken state; it's not hurting anyone for now.

The bug is that (1) querying a constant whose value is a single-valued tensor in the original model automatically converts that tensor to `True` or `False` depending on whether its single value is nonzero or zero, and (2) querying a multiple-valued tensor will crash.

I'll follow up on the bug fix in an upcoming diff stack.

Reviewed By: wtaha

Differential Revision: D28886660

fbshipit-source-id: c46b0c4750a81047382ef82609db6db9d0ac0634