Fix code and graph generation for tensor constants (#878)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/878

When we generate a BMG graph from a model, or generate the Python code to do so, we were doing it wrong for a certain set of tensor values; the bug was mostly unobservable until recently, when I added ability to query directly on a constant value.

We no longer crash when generating a query on a multi-value tensor, and no longer automatically convert the tensor to bool when querying a single-value tensor.

The C++ code generation still looks wrong but we can fix that in a later diff.

There is still the lingering problem that `add_constant(tensor(2.5))` does not do what you think it does in Python because the python bindings turn that into `add_constant(True)`.  I'll fix that in an upcoming diff.

Reviewed By: wtaha

Differential Revision: D28911908

fbshipit-source-id: de80b14183f598150fe6858183d7dec25b4d08a9