Accumulate non-numeric constants into graph, reject them later
Summary:
While executing the model, the runtime adds nodes to the graph accumulator every time an operation has at least one stochastic operand (that is, an operand that is already a graph node.)  What should we do if the operation also has a non-stochastic "constant" operand?

Previously, we checked to see if the value could be represented in BMG -- if the value was a bool, int, real, single-valued tensor, or tensor with dimensionality<=2> we'd accept it, and if not then we'd crash during graph accumulation.

This is problematic for a number of reasons.

* BMG might not be the only backend we want to support; crashing early just because we cannot represent a thing in BMG makes supporting other backends harder.

* We might not want to generate BMG at all; we might just want to do an analysis of the stochastic operations in the model for some other reason.

* There may be transformations we can do on unsupported constant values to obtain an equivalent graph that is supported in BMG, but we cannot do that if we've already crashed.

* The error reporting experience is poor for the user.

* The error reporting code should, when possible, be isolated to analysis passes which detect specific problems, repair them if possible, and give sensible, understandable error messages if the problem cannot be fixed.

I've updated the graph accumulation code to always put a constant value into an UntypedConstantNode when it is used as an operand to an operator with at least one stochastic operand, and not check whether it is a supported-by-BMG value until a later error detection and repair pass.

This led to a number of knock-on effects on type analysis, but the summary is: we previously took advantage of the invariant that every UntypedConstantNode could be converted to a node with a supported type in the BMG type system. That's no longer the case, so now UntypedConstantNodes can be judged to be Untypeable.

Reviewed By: wtaha

Differential Revision: D28692541

fbshipit-source-id: f3b916ff2e74fa63809f49cfbda87c15ec19bafd