Test to provide a checked graph representation for an n-schools example
Summary:
To support an external collaboration we are formalising some example n-schools models, and in particular, the different graph representations related these models. In particular, we expect to share graphs for both handwritten and Beanstalk generated examples of the same model.

Note: Several starting points to additional models have for future explorations are also included in this diff

{F539446594}

Reviewed By: ericlippert

Differential Revision: D27315653

fbshipit-source-id: 02bac6d305c5f5a46d1f203c496a3e53fc57465b