Add error messages for remaining unsupported comparison operators (#881)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/881

BMG does not support any stochastic comparison operators.  We already give sensible error message for uses of `< > <= >= != ==` with one or two stochastic operands but we did not give an error for unsupported uses of `is, is not, in, not in` which Python classifies as comparisons.

We now accumulate these operations into the graph and then report in the unsupported graph node pass if such a node is in the graph. This was, if we do eventually support any of these operations in BMG we can do so easily.

Reviewed By: wtaha

Differential Revision: D28430535

fbshipit-source-id: 1e4bdf82e172697b10fe1e7fda087b95c652dd2a