Remove/fix incorrect assertions (#868)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/868

Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/868

Assertions are currently disabled in BMG, which means that as code has been edited, errors have crept into existing assertions that were not flagged by the compiler.  I mean to eventually turn assertions back on, so the first step is to eliminate or fix incorrect assertions. See comments below for details.

Reviewed By: wtaha

Differential Revision: D29040345

fbshipit-source-id: f507c965a70244bec5ae300b68bec3670cc2ecfb