Correct sample generator for half_normal
Summary: This diff fixes the sample generator in the code so that it is a proper sample generator for half_normal rather than normal. This is the third diff in a series of five diffs planned to complete the implementation of the half_normal distribution (see D28612073). The tests are also updated accordingly, and in particular, the mean and variance are adjusted per definitions in wikipedia page https://en.wikipedia.org/wiki/Half-normal_distribution

Reviewed By: yucenli

Differential Revision: D29155563

fbshipit-source-id: 8058d09d8e1ad874462f221b8bc143a524d53c70