Setting mean to zero and updating tests
Summary: This is the first step towards turning this copy of the implementation of normal distribution into an actual implementation of the half normal distribution. In particular, it is the first of five steps described in a previous diff (D28612073). Compared to normal, half normal takes only one parameter rather than two, as the mean is implicilty taken to be 0. In this diff, we set the mean to zero and update the tests to make sure that they continue to pass. In the next diff, we will remove this parameter from the implementation, and update the tests accordingly.

Reviewed By: yucenli

Differential Revision: D29050167

fbshipit-source-id: c96ade7c1719c9e0808242a65a024527bf3fb616