Comments about how the (copied and renamed) normal distribution code works
Summary:
This diff simply adds some comments about parts of this code that needed clarification. Earlier versions also included questions that were kindly answered by Lily and Rodrigo. Going forward, the plan is to implement half_normal in the following steps: 1) Set the mean parameters on all instances to zero and make sure that all tests are updated accordingly, 2) Removing the mean parameter, 3) updating the sampler, 4) updating the CDF, 5) updating the type of the sample from REAL to POS_REAL.

P.S. Most of the TODO items in the current file are expected to be addressed through the five upcoming diffs outlined above.

Reviewed By: yucenli

Differential Revision: D28612073

fbshipit-source-id: e9bba4bd5a2424bd41d7ece0efddd51342b66ebb