Basic end to end test for Beanstalk for a simple normal distribution example
Summary: To facilitate testing of extensions of the compiler with support for new distributions, it is useful to have end to end tests for individual distributions. This diff provides a test for the normal distribution, which will be used as a template for a test for the newly added half_normal distribution in an upcoming diff.

Reviewed By: ericlippert

Differential Revision: D29319961

fbshipit-source-id: fb93e0bd519efaabe4e03c95bcdd20fd3511d59e