Converting sample type for half_normal to POS_REAL
Summary:
This diff changes the type of samples from half_normal from REAL to POS_REAL, and is the last of five steps to convert a copy of the code for normal to half_normal (see D28612073) for details.

This completes the addition of the functional of half_normal. I intend to make one final pass over the code in these two files once the tutorial example that motivated this addition compiles and runs.

Reviewed By: yucenli

Differential Revision: D29232141

fbshipit-source-id: 7f4a1980671e7ee04c03d12f7b5488cdf14fcf19