Updated logprob for half_normal
Summary: This diff corrects the definition of the logarithm of the probability distribution to be that of the half_normal distribution rather than that of the normal distribution. This is the fourth step in a series of five steps to complete the implementation of the half_normal distribution (see D28612073).

Reviewed By: yucenli

Differential Revision: D29155895

fbshipit-source-id: 31cd775607ddec7be64174bb428390644e6b7cc3