Starting to implement half_normal (#846)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/846

This diff begins the implementation of the half_normal distribution in BMG. This diff simply makes a copy of the files for the normal distribution, which will serve as a template for the half_normal implementation. This diff only creates a minimally adapted copy of those template files so that future diffs that modify specific parts of those files are easier to read. The changes made are the minimal ones made to make the test (for the copy of Normal, not an actual Half_Normal) work.

Reviewed By: yucenli

Differential Revision: D28496433

fbshipit-source-id: 44a5b9c01ad15c4cda08f1a97a69830bb84ae1f1