Template to test original and conjugate prior-based transformed model (#887)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/887

This is the first diff for my internship project. The plan is to implement conjugate prior-based transformations. This diff adds tests to check that original and transformed models have equivalent output. Note, here I chose a specific case of transformation, namely beta-binomial.

Given a model like `HeadsRate`, with Bernoulli likelihood and Beta prior, we want Beanstalk compiler to identify that it is an instance of conjugate prior and produce a transformed model that looks like `HeadsRateTransformed`.

First, the transformation would identify that the queried parameter `theta()` has a Beta-Binomial conjugate relationship. It  analytically computes the posterior parameters (specifically for our example it would be `Beta(alpha + sum y_i, beta + n - sum y_i)`) , based on the prior parameters and observations. Now it inserts a new random variable function, `theta_transformed()` . Now we query this transformed function and further, the observations are set to be empty. We modify calls to inference with the updated queries and observations. The idea is that the resulting model samples directly from the transformed analytical posterior.

Currently this test checks (using KS) if these two models produce samples from the same distribution. Once we have implemented the transformation, we could additionally check that the graphs produced from `HeadsRate` and `HeadsRateTransformed` are syntactically same.

Reviewed By: ericlippert, horizon-blue

Differential Revision: D29430975

fbshipit-source-id: 7b03aee18a9eca47bcebc8481d1eeb43926a709f