Include target_node as parameter of `Node::gradient_log_prob`.
Summary:
Modified `graph::Node::gradient_log_prob` to receive target node as parameter. This makes things conceptually simpler, dispensing with the need of using the gradient value as an indirect indicator of whether the gradient is being computed on the target node itself or some other node.

This was also an attempt at making it unnecessary to clear the gradients after each iteration, since the previous version of this method relied on all gradients other than the target node's being zero. The diff does indeed remove this unnecessary dependency, but some other code must be depending on gradients being cleared since tests fail if we remove gradient clearing. It is worthwhile keeping a eye on gradient computations from now on to see if we identify this other dependency, and once we do we can hopefully not need to clear gradients anymore.

Reviewed By: yucenli

Differential Revision: D29341532

fbshipit-source-id: f44663f58b63f73980caf0b6307d40eb49edb316