Set unconstrained value of simplex to broadcast matrix (#897)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/897

To avoid allocating memory unnecessarily we set the unconstrained value of a sample node to either an atomic real or a 1x0 array of reals. Consider however the Dirichlet distribution, where the sample type is a simplex; a ValueType with variable type COL_SIMPLEX_MATRIX and atomic type REAL violates the invariants of ValueType.  (A simplex is required to have elements that are probabilities.)

We avoid the problem here by setting the unconstrained value to a broadcast matrix of reals in this scenario.

Reviewed By: yucenli

Differential Revision: D29599594

fbshipit-source-id: dbe8c3b3df8ebb108165b661c3074a39c0f77858