Mark private members as private
Summary:
In reviewing code in BMG I noticed that there were a few members of Graph that really should be marked as private because they are only used internally and expose internal implementation details. I've marked them as private.

There was one where we cannot mark it as private even though it should be; a test depends on it being public. I've left a note to fix that later.

Reviewed By: wtaha

Differential Revision: D29072254

fbshipit-source-id: 6d7fdde34925ec1285fe61c47c1eec3ded3a592b