Multi-process for global inference
Summary:
This diff adds an option to run global inference in parallel (`run_in_parallel`) together with an option to set the seed for a run. The name of the parameter is chosen to be consistent with the single site inference method.

Because Python uses Pickle to transfer data between process, and...
- When an object gets pickle and then unpickle in another process, this effectively creates a copy of the object *at a different address with a different `id`*
- A random variable define as an method receives the object as its first argument (i.e. the `self` argument)
- Our BM uses a function and its arguments to identify a random variable

This means that we can't just pass a mapping from `RVIdentifier` to `torch.Tensor` back to the main process in general because the hashs of the `RVIdentifier`s are different. In single site inference, we cast [`RVIdentifier` to string as a workaround](https://www.internalfb.com/code/fbsource/[5b473e1c6f3e47990da858028181b26577bc0e7a]/fbcode/beanmachine/beanmachine/ppl/inference/abstract_infer.py?lines=203%2C270) (Though on a second thought, it's quite possible to run into name conflict).

So in this diff, instead of casting to string, I use the order of the variables to determine the mapping from query to values.

A few issues that might need to follow up:
- The progress bar doesn't work in parallel runs (at least not when I tested it in Bento)
- PyTorch [prompt me to use `spawn` instead of `fork`](https://github.com/pytorch/pytorch/wiki/Autograd-and-Fork), but when I do so, TestX times out and Bento runs indefinitely. I haven't figure out what's the best way to handle this so I didn't do anything :/

(See N973238 for some illustration)

Differential Revision: D29629659

fbshipit-source-id: 7fe56b89c77b41fad79e71d036c9c044c37febd8