Fix observe error detection (#901)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/901

The observe() APIs in BMG did not correctly validate their arguments; rather, they assumed, for instance, that when given an observation of a 2x3 matrix of doubles, that the observed quantity was in fact a 2x3 matrix of some type with storage type double, and then violated a bunch of type system invariants along the way.

We now check early on that an observation matches the sample being observed.

We do not yet do range checks; that is, an observation of a probability must be a double, but we do not check that it is between 0 and 1.  I might add that in a later diff.

Reviewed By: yucenli

Differential Revision: D29073748

fbshipit-source-id: 940ed50969dac7f3fce40466c88f02a353af00d0