Selectively disable optimizations when generating BMG (#889)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/889

While trying to benchmark performance of binary vs multiary addition, I came across the following:

In Beanstalk error-detection (currently 6) and optimizations (currently 1) together are called fixer types. There is a flag named  `after_transform` in `BMGInference().to_dot(...)` that either enables/disables all of these fixer types together. This flag was introduced for debugging purpose to answer one of the two scenarios. (1) what does the graph look like immediately after graph accumulation is done (after_transform = False)  and (2) what graph are we actually going to send to BMG (after_transform = True).

There was no flag to selectively just enable/disable optimizations. So I added code to skip optimizations that have been passed as skip.  We expose the optimization toggle as a set of strings (which are names of the optimization fixer type classes) that a user could input to the `inference` and `to_dot` method of `BMGInference`. We also maintain a `default_skip_optimizations` set (currently empty) to make sure by default we carry out all optimizations.

Alternatively, we could separate the `fixer_types` into two lists and separate them. However, performance reports require the ordering of these to be same. For now to make minimal changes and to prevent report generation tests from breaking, I have left the `fixer_types` as a single list.

I added two test cases to check the correctness of this change. First we check if the graphs without and with optimizations are as expected. Further, we check if the KS distance between samples from optimized and un-optimized models are from same distribution.

Graph when multiary addition optimization is toggled off

{F629466640}

Graph when multiary addition optimization is toggled on
{F629466723}

Reviewed By: ericlippert

Differential Revision: D29504284

fbshipit-source-id: 8bc71c1384ec2d42e0e08db63e28634b7226bd30