Added node types and edge requirements for multiary multiplication
Summary:
Added edge requirements and node type for the added multiary multiplication node.

`Node Type` : The type of this new node should be the same as the original multiplication node with but extended to support > 2 nodes per operation. Hence modified  `_type_multiplication` to first compute the supremum operator type of all input nodes. The remaining logic is the same as binary multiplication.

`Edge Requirement`: The edge requirement for this node with `n` inputs is a list of `n` requirements, one corresponding to each edge. Hence modified `_requirements_multiplication` to return a list of `n` requirements instead of two.

Reviewed By: ericlippert

Differential Revision: D29563905

fbshipit-source-id: 4d22a1d29d9514f4c89b69245bb66469394e3749