Implement multiary multiplication node in graph builder
Summary:
Added a multary multiplication node in the graph builder. There is no functionality associated with it yet.
This will be used in a proposed optimization to turn ((a * b) * c) * d into a single multiary *(a, b, c, d) node, in an upcoming diff. Type requirements check will also be done in an upcoming diff.

Differential Revision: D29559757

fbshipit-source-id: 7c6f59a401b111159b934d692db41d793ae29120