Rewrite binary multiplication into multiary multiplication
Summary:
BMG supports a multiary multiplication node that can take many operands. By using it we turn n add nodes with a total of 2n edges into a single multiary multiplication node with a total of n edges. (similar to the multiary sum node)

Determining exactly which * nodes to turn into multiary multiplication follows the same heuristic as multiary addition. For the heuristic, see the comment in `MultiaryOperatorFixer
._needs_fixing`.

For this optimization I have created a new class `MultiaryMultiplicationFixer`. This choice was made specifically to enable selective toggling of optimizations. We could consider creating a base class `MultiaryOperatorFixer` and adding `MultiaryAdditionFixer`. For now, to make minimal changes I have left `MultiaryOperatorFixer` as is.

The new test case shows a simple example of how the rewriter works. We start with this model, which computes multiplication of six samples from a normal:

{F630277097}

The rewriter detects that two of those multiplications can be rewritten to lower the edge count and node count of the graph:

{F630277253}

The resulting graph has two fewer nodes and two fewer edges but does exactly the same amount of arithmetic.

Reviewed By: ericlippert

Differential Revision: D29568019

fbshipit-source-id: 32a637b12a60b595255d4e30590516960900c847