Added node types and edge requirements for multiary multiplication (#906)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/906

Added edge requirements and node type for the added multiary multiplication node.

`Node Type` : The type of this new node should be the same as the original multiplication node with but extended to support > 2 nodes per operation. Hence modified  `_type_multiplication` to first compute the supremum operator type of all input nodes. The remaining logic is the same as binary multiplication.

`Edge Requirement`: The edge requirement for this node with `n` inputs is a list of `n` requirements, one corresponding to each edge. Hence modified `_requirements_multiplication` to return a list of `n` requirements instead of two.

Reviewed By: ericlippert

Differential Revision: D29563905

fbshipit-source-id: e161d5071300a1b7597cc7343533b46f172fec88